Free play (#12)
Free Play mode is now available. 
In Free Play there is no target to match, and the potato has a number of different slots. Clicking a component reveals the slots, and clicking any slot places the component in that slot. 
Components can be removed from the potato by clicking them, and the potato can be made to dance with a button.